tokenlist.json
XSah Coin is a community-powered Solana token born from the spirit of the Kurdish mountains. It represents survival, freedom, and cultural pride — inspired by the image of a mountain ram fighting against nature, yet never bowing.

This token is not just a meme. It is a message.

Our goal is to build a decentralized and resilient community around values of resistance and self-determination.

XSah means: FREE LIKE A PEOPLE 🐏